refactor(WinningNumbers): 당첨 번호를 SixNumbers와 BonusNumber로 분리
